+++ CHANGELOG.md
@@ -61,6 +61,7 @@
 - Added support for `Series.dt.is_quarter_start` and `Series.dt.is_quarter_end`.
 - Added support for `Series.argmax` and `Series.argmin`.
 - Added support for `Series.dt.is_leap_year`.
+- Added support for `Index.__repr__`.
 
 #### Improvements
 - Removed the public preview warning message upon importing Snowpark pandas.